fix(chat-room): dispose the controller on unmount, and run the browser paths cross-engine
**Leaving the page mid-stream leaked a run.** Destroying `<Chat>` does
not destroy the controller behind it, so the run kept going, its frames
kept arriving for a component that was gone, and the provider request
stayed open and billed. `session.dispose()` on `onDestroy` stops the
active run and releases the controller's subscriptions.

**The browser paths now run cross-engine**, which is what the review
asked for and what my own corrected comment said should happen. Both
tests that drive a real engine path — the create flow's `fetch` plus
reload, and the incremental test's `ReadableStream` through the session
controller — moved into their own spec so the shards do not also triple
the `request`-fixture tests, which drive a Node-side HTTP client.

Placed by measurement rather than by eye, which this repository has been
bitten by twice: `--list` per project read 49/48/62/51/50, so the two
tests go to `webkit-2` (48 → 50) and not to `webkit-3`, already at the 62
the ceiling note describes. Re-measured after: 49/50/62/51/50.

Both pass under WebKit, so the streaming path now has coverage in an
engine other than Chromium for the first time.
